fix: security configuration, open api docs
