+++ src/main/java/org/example/learningsystem/core/web/swagger/CloudOpenApiConfiguration.java
@@ -31,6 +31,8 @@ public class CloudOpenApiConfiguration extends OpenApiConfiguration {
     public OpenAPI openAPI() {
         return super.openAPI()
                 .addSecurityItem(new SecurityRequirement()
+                        .addList(BASIC_AUTH_SCHEME))
+                .addSecurityItem(new SecurityRequirement()
                         .addList(BEARER_AUTH_SCHEME))
                 .components(new Components()
                         .addSecuritySchemes(BASIC_AUTH_SCHEME, basicAuthSecurityScheme())
