+++ src/main/java/org/example/learningsystem/multitenancy/subscription/controller/SubscriptionController.java
@@ -1,5 +1,9 @@
 package org.example.learningsystem.multitenancy.subscription.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.learningsystem.multitenancy.subscription.dto.ServiceInfoDto;
@@ -23,11 +27,17 @@ import static org.springframework.http.HttpStatus.NO_CONTENT;
 @RequestMapping("/api/v1/subscriptions")
 @RequiredArgsConstructor
 @Slf4j
+@Tag(name = "Subscription Controller")
 public class SubscriptionController {
 
     private final SubscriptionService subscriptionService;
 
     @PutMapping("/tenants/{tenantId}")
+    @Operation(summary = "Subscribe to application")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Course was created"),
+            @ApiResponse(responseCode = "400", description = "Invalid body")
+    })
     public String subscribe(
             @PathVariable("tenantId") String tenantId, @RequestBody SubscriptionRequestDto subscription) {
         log.info("Subscription request [tenantId = {}, subdomain = {}]", tenantId, subscription.subscribedSubdomain());
@@ -36,6 +46,8 @@ public class SubscriptionController {
 
     @DeleteMapping("/tenants/{tenantId}")
     @ResponseStatus(NO_CONTENT)
+    @Operation(summary = "Unsubscribe from application")
+    @ApiResponse(responseCode = "204", description = "Unsubscription was successful")
     public void unsubscribe(
             @PathVariable("tenantId") String tenantId, @RequestBody SubscriptionRequestDto subscription) {
         log.info("Delete subscription request [tenantId = {}, subdomain = {}]", tenantId, subscription.subscribedSubdomain());
@@ -43,6 +55,7 @@ public class SubscriptionController {
     }
 
     @GetMapping("/dependencies")
+    @Operation(summary = "Get application dependencies")
     public List<ServiceInfoDto> getDependencies(@RequestParam("tenantId") String tenantId) {
         log.info("Dependencies request [tenantId = {}]", tenantId);
         var dependencies = subscriptionService.getDependencies();
